requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.8
    # via realtime
aiosignal==1.3.2
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anthropic==0.52.2
    # via pydantic-ai-slim
anyio==4.9.0
    # via
    #   anthropic
    #   google-genai
    #   groq
    #   httpx
    #   mcp
    #   openai
    #   pydantic-evals
    #   sse-starlette
    #   starlette
argcomplete==3.6.2
    # via pydantic-ai-slim
attrs==25.3.0
    # via aiohttp
black==25.1.0
    # via -r requirements.in
boto3==1.38.29
    # via pydantic-ai-slim
botocore==1.38.29
    # via
    #   boto3
    #   s3transfer
cachetools==5.5.2
    # via google-auth
certifi==2025.4.26
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via
    #   black
    #   uvicorn
cohere==5.15.0
    # via pydantic-ai-slim
colorama==0.4.6
    # via
    #   click
    #   griffe
    #   pytest
    #   tqdm
deprecated==1.2.18
    # via
    #   opentelemetry-api
    #   opentelemetry-exporter-otlp-proto-http
    #   opentelemetry-semantic-conventions
deprecation==2.1.0
    # via postgrest
distro==1.9.0
    # via
    #   anthropic
    #   groq
    #   openai
ecdsa==0.19.1
    # via python-jose
eval-type-backport==0.2.2
    # via
    #   mistralai
    #   pydantic-ai-slim
executing==2.2.0
    # via logfire
fasta2a==0.2.14
    # via pydantic-ai-slim
fastapi==0.115.12
    # via -r requirements.in
fastavro==1.11.1
    # via cohere
filelock==3.18.0
    # via huggingface-hub
frozenlist==1.6.2
    # via
    #   aiohttp
    #   aiosignal
fsspec==2025.5.1
    # via huggingface-hub
google-auth==2.40.2
    # via
    #   google-genai
    #   pydantic-ai-slim
google-genai==1.18.0
    # via pydantic-ai-slim
googleapis-common-protos==1.70.0
    # via opentelemetry-exporter-otlp-proto-http
gotrue==2.12.0
    # via supabase
griffe==1.7.3
    # via pydantic-ai-slim
groq==0.26.0
    # via pydantic-ai-slim
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
h2==4.2.0
    # via httpx
hpack==4.1.0
    # via h2
httpcore==1.0.9
    # via httpx
httpx[http2]==0.28.1
    # via
    #   anthropic
    #   cohere
    #   google-genai
    #   gotrue
    #   groq
    #   mcp
    #   mistralai
    #   openai
    #   postgrest
    #   pydantic-ai-slim
    #   pydantic-graph
    #   storage3
    #   supabase
    #   supafunc
httpx-sse==0.4.0
    # via
    #   cohere
    #   mcp
huggingface-hub==0.32.4
    # via tokenizers
hyperframe==6.1.0
    # via h2
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
importlib-metadata==8.6.1
    # via opentelemetry-api
iniconfig==2.1.0
    # via pytest
isort==6.0.1
    # via -r requirements.in
jiter==0.10.0
    # via
    #   anthropic
    #   openai
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
logfire==3.17.0
    # via -r requirements.in
logfire-api==3.17.0
    # via
    #   pydantic-evals
    #   pydantic-graph
markdown-it-py==3.0.0
    # via rich
mcp==1.9.2
    # via pydantic-ai-slim
mdurl==0.1.2
    # via markdown-it-py
mistralai==1.8.1
    # via pydantic-ai-slim
multidict==6.4.4
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.1.0
    # via black
openai==1.84.0
    # via pydantic-ai-slim
opentelemetry-api==1.33.1
    # via
    #   fasta2a
    #   opentelemetry-exporter-otlp-proto-http
    #   opentelemetry-instrumentation
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
    #   pydantic-ai-slim
opentelemetry-exporter-otlp-proto-common==1.33.1
    # via opentelemetry-exporter-otlp-proto-http
opentelemetry-exporter-otlp-proto-http==1.33.1
    # via logfire
opentelemetry-instrumentation==0.54b1
    # via logfire
opentelemetry-proto==1.33.1
    # via
    #   opentelemetry-exporter-otlp-proto-common
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-sdk==1.33.1
    # via
    #   logfire
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-semantic-conventions==0.54b1
    # via
    #   opentelemetry-instrumentation
    #   opentelemetry-sdk
packaging==25.0
    # via
    #   black
    #   deprecation
    #   huggingface-hub
    #   opentelemetry-instrumentation
    #   pytest
pathspec==0.12.1
    # via black
platformdirs==4.3.8
    # via black
pluggy==1.6.0
    # via pytest
postgrest==1.0.2
    # via supabase
prompt-toolkit==3.0.51
    # via pydantic-ai-slim
propcache==0.3.1
    # via
    #   aiohttp
    #   yarl
protobuf==5.29.5
    # via
    #   googleapis-common-protos
    #   logfire
    #   opentelemetry-proto
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   python-jose
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.11.5
    # via
    #   anthropic
    #   cohere
    #   fasta2a
    #   fastapi
    #   google-genai
    #   gotrue
    #   groq
    #   mcp
    #   mistralai
    #   openai
    #   postgrest
    #   pydantic-ai-slim
    #   pydantic-evals
    #   pydantic-graph
    #   pydantic-settings
pydantic-ai==0.2.14
    # via -r requirements.in
pydantic-ai-slim[a2a,anthropic,bedrock,cli,cohere,evals,google,groq,mcp,mistral,openai,vertexai]==0.2.14
    # via
    #   pydantic-ai
    #   pydantic-evals
pydantic-core==2.33.2
    # via
    #   cohere
    #   pydantic
pydantic-evals==0.2.14
    # via pydantic-ai-slim
pydantic-graph==0.2.14
    # via pydantic-ai-slim
pydantic-settings==2.9.1
    # via mcp
pygments==2.19.1
    # via
    #   pytest
    #   rich
pyjwt==2.10.1
    # via gotrue
pytest==8.4.0
    # via
    #   pytest-asyncio
    #   pytest-mock
pytest-asyncio==1.0.0
    # via -r requirements.in
pytest-mock==3.14.1
    # via gotrue
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   mistralai
    #   realtime
    #   storage3
python-dotenv==1.1.0
    # via pydantic-settings
python-jose==3.5.0
    # via -r requirements.in
python-multipart==0.0.20
    # via mcp
pyyaml==6.0.2
    # via
    #   huggingface-hub
    #   pydantic-evals
realtime==2.4.3
    # via supabase
requests==2.32.3
    # via
    #   cohere
    #   google-genai
    #   huggingface-hub
    #   opentelemetry-exporter-otlp-proto-http
    #   pydantic-ai-slim
rich==14.0.0
    # via
    #   logfire
    #   pydantic-ai-slim
    #   pydantic-evals
rsa==4.9.1
    # via
    #   google-auth
    #   python-jose
ruff==0.11.12
    # via -r requirements.in
s3transfer==0.13.0
    # via boto3
six==1.17.0
    # via
    #   ecdsa
    #   python-dateutil
sniffio==1.3.1
    # via
    #   anthropic
    #   anyio
    #   groq
    #   openai
sse-starlette==2.3.6
    # via mcp
starlette==0.46.2
    # via
    #   fasta2a
    #   fastapi
    #   mcp
storage3==0.11.3
    # via supabase
strenum==0.4.15
    # via supafunc
supabase==2.15.2
    # via -r requirements.in
supafunc==0.9.4
    # via supabase
tokenizers==0.21.1
    # via cohere
tqdm==4.67.1
    # via
    #   huggingface-hub
    #   openai
types-requests==2.32.0.20250602
    # via cohere
typing-extensions==4.14.0
    # via
    #   anthropic
    #   anyio
    #   cohere
    #   fastapi
    #   google-genai
    #   groq
    #   huggingface-hub
    #   logfire
    #   openai
    #   opentelemetry-sdk
    #   pydantic
    #   pydantic-core
    #   realtime
    #   typing-inspection
typing-inspection==0.4.1
    # via
    #   mistralai
    #   pydantic
    #   pydantic-ai-slim
    #   pydantic-graph
    #   pydantic-settings
urllib3==2.4.0
    # via
    #   botocore
    #   requests
    #   types-requests
uvicorn==0.34.3
    # via mcp
wcwidth==0.2.13
    # via prompt-toolkit
websockets==14.2
    # via
    #   google-genai
    #   realtime
wrapt==1.17.2
    # via
    #   deprecated
    #   opentelemetry-instrumentation
yarl==1.20.0
    # via aiohttp
zipp==3.22.0
    # via importlib-metadata
